refactor(update lock status from backend): move to its own file
