+++ src/shared/locked-status/index.js
@@ -1,8 +1,6 @@
-export {
-    useCheckLockStatus,
-    updateLockStatusFromBackend,
-} from './use-check-lock-status.js'
 export { LockedContext } from './locked-info-context.js'
 export { LockedProvider } from './locked-info-provider.js'
 export { LockedStates } from './locked-states.js'
+export { useCheckLockStatus } from './use-check-lock-status.js'
 export { useLockedContext } from './use-locked-context.js'
+export { updateLockStatusFromBackend } from './update-lock-status-from-backend.js'
